gpu: nvgpu: regops: u32 num_ops for exec_regops
The exec_regops() API was using a u64 for the num_ops parameter. The
lower level APIs used by exec_regops() expect u32s for this value.
Update the interface to use u32.

This eliminates MISRA Rule 10.3 violations for assignment of objects of
different essential or narrower types.

JIRA: NVGPU-3023

Reviewed-on: https://git-master.nvidia.com/r/2084207
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/regops/regops.c
@@ -86,7 +86,7 @@ static bool validate_reg_ops(struct gk20a *g,
 int exec_regops_gk20a(struct gk20a *g,
 		      struct channel_gk20a *ch,
 		      struct nvgpu_dbg_reg_op *ops,
-		      u64 num_ops,
+		      u32 num_ops,
 		      bool is_profiler,
 		      bool *is_current_ctx)
 {
