+++ drivers/gpu/nvgpu/include/nvgpu/gk20a.h
@@ -1479,7 +1479,7 @@ struct gpu_ops {
 		int (*exec_regops)(struct gk20a *g,
 			    struct channel_gk20a *ch,
 			    struct nvgpu_dbg_reg_op *ops,
-			    u64 num_ops,
+			    u32 num_ops,
 			    bool is_profiler,
 			    bool *is_current_ctx);
 		const struct regop_offset_range* (
